[feat] 페이지 별 공통기능 (api 값 불러오기)를 커스텀훅으로 별도 파일 구분 및 적용
